Created confirmation password validation. refactored packets structure closes #18
