+++ src/main/java/pl/olpinski/stickynotes/Config.java
@@ -3,12 +3,9 @@ package pl.olpinski.stickynotes;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Scope;
-import org.springframework.security.authentication.AuthenticationProvider;
-import org.springframework.security.core.Authentication;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
-import pl.olpinski.stickynotes.service.UserService;
-import pl.olpinski.stickynotes.web.security.provider.DatabaseAuthenticationProvider;
+
 
 @Configuration
 public class Config {
@@ -18,4 +15,5 @@ public class Config {
     public PasswordEncoder passwordEncoder() {
         return new BCryptPasswordEncoder();
     }
+
 }
