+++ src/main/java/pl/olpinski/stickynotes/data/converter/UserConverter.java
@@ -3,10 +3,9 @@ package pl.olpinski.stickynotes.data.converter;
 import org.springframework.stereotype.Component;
 import pl.olpinski.stickynotes.data.entity.Note;
 import pl.olpinski.stickynotes.data.entity.User;
-import pl.olpinski.stickynotes.data.entity.UserStatus;
-import pl.olpinski.stickynotes.dto.NewUserDto;
-import pl.olpinski.stickynotes.dto.NoteDto;
-import pl.olpinski.stickynotes.dto.UserDto;
+import pl.olpinski.stickynotes.data.dto.NewUserDto;
+import pl.olpinski.stickynotes.data.dto.NoteDto;
+import pl.olpinski.stickynotes.data.dto.UserDto;
 
 import java.util.Set;
 import java.util.stream.Collectors;
