--- src/main/java/pl/olpinski/stickynotes/dto/NoteDto.java
@@ -1,43 +1,0 @@
-package pl.olpinski.stickynotes.dto;
-
-public class NoteDto {
-
-    private Long id;
-    private Long userId;
-    private String title;
-    private String content;
-
-
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public Long getUserId() {
-        return userId;
-    }
-
-    public void setUserId(Long userId) {
-        this.userId = userId;
-    }
-
-    public String getTitle() {
-        return title;
-    }
-
-    public void setTitle(String title) {
-        this.title = title;
-    }
-
-    public String getContent() {
-        return content;
-    }
-
-    public void setContent(String content) {
-        this.content = content;
-    }
-}
