--- src/main/java/pl/olpinski/stickynotes/dto/UserDto.java
@@ -1,55 +1,0 @@
-package pl.olpinski.stickynotes.dto;
-
-import java.util.Set;
-
-public class UserDto {
-
-    private Long id;
-
-    private String login;
-    private String password;
-    private String mail;
-
-    private Set<NoteDto> notes;
-
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getLogin() {
-        return login;
-    }
-
-    public void setLogin(String login) {
-        this.login = login;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-
-    public String getMail() {
-        return mail;
-    }
-
-    public void setMail(String mail) {
-        this.mail = mail;
-    }
-
-    public Set<NoteDto> getNotes() {
-        return notes;
-    }
-
-    public void setNotes(Set<NoteDto> notes) {
-        this.notes = notes;
-    }
-}
