+++ src/main/java/pl/olpinski/stickynotes/service/NoteService.java
@@ -1,8 +1,8 @@
 package pl.olpinski.stickynotes.service;
 
 import pl.olpinski.stickynotes.data.entity.Note;
-import pl.olpinski.stickynotes.dto.NewNoteDto;
-import pl.olpinski.stickynotes.dto.NoteDto;
+import pl.olpinski.stickynotes.data.dto.NewNoteDto;
+import pl.olpinski.stickynotes.data.dto.NoteDto;
 
 public interface NoteService {
