+++ src/main/java/pl/olpinski/stickynotes/service/impl/UserServiceImpl.java
@@ -6,8 +6,8 @@ import org.springframework.stereotype.Service;
 import pl.olpinski.stickynotes.data.converter.UserConverter;
 import pl.olpinski.stickynotes.data.entity.User;
 import pl.olpinski.stickynotes.data.entity.UserStatus;
-import pl.olpinski.stickynotes.dto.NewUserDto;
-import pl.olpinski.stickynotes.dto.UserDto;
+import pl.olpinski.stickynotes.data.dto.NewUserDto;
+import pl.olpinski.stickynotes.data.dto.UserDto;
 import pl.olpinski.stickynotes.data.repository.UserRepository;
 import pl.olpinski.stickynotes.service.MailService;
 import pl.olpinski.stickynotes.service.UserService;
