+++ src/main/java/pl/olpinski/stickynotes/web/controller/HomeController.java
@@ -1,0 +1,28 @@
+package pl.olpinski.stickynotes.web.controller;
+
+import org.springframework.security.core.Authentication;
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.GetMapping;
+import pl.olpinski.stickynotes.data.dto.UserDto;
+import pl.olpinski.stickynotes.service.UserService;
+
+@Controller
+public class HomeController {
+
+    UserService userService;
+
+    public HomeController(UserService userService) {
+        this.userService = userService;
+    }
+
+    @GetMapping({"/notes", "", "/"})
+    public String user(Model model, Authentication authentication){
+
+        Long id = (Long) authentication.getPrincipal();
+        UserDto userDto = userService.findUserById(id);
+        model.addAttribute("user", userDto);
+        return "notes";
+    }
+
+}
