+++ src/main/java/pl/olpinski/stickynotes/web/controller/NoteController.java
@@ -6,9 +6,8 @@ import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.ModelAndView;
 import pl.olpinski.stickynotes.data.entity.Note;
-import pl.olpinski.stickynotes.dto.NewNoteDto;
-import pl.olpinski.stickynotes.dto.NoteDto;
-import pl.olpinski.stickynotes.dto.UserDto;
+import pl.olpinski.stickynotes.data.dto.NewNoteDto;
+import pl.olpinski.stickynotes.data.dto.NoteDto;
 import pl.olpinski.stickynotes.service.NoteService;
 import pl.olpinski.stickynotes.service.UserService;
 
@@ -25,14 +24,6 @@ public class NoteController {
     }
 
 
-    @GetMapping({"/notes", "", "/"})
-    public String user(Model model, Authentication authentication){
-
-        Long id = (Long) authentication.getPrincipal();
-        UserDto userDto = userService.findUserById(id);
-        model.addAttribute("user", userDto);
-        return "notes";
-    }
 
     @GetMapping("/{id}")
     public String note(@PathVariable Long id, Model model){
