+++ src/main/java/pl/olpinski/stickynotes/web/validation/UniqueMail.java
@@ -1,15 +1,15 @@
-package pl.olpinski.stickynotes.validation;
+package pl.olpinski.stickynotes.web.validation;
 
 import javax.validation.Constraint;
 import javax.validation.Payload;
 import java.lang.annotation.*;
 
 @Documented
-@Constraint(validatedBy = UniqueLoginValidator.class)
+@Constraint(validatedBy = UniqueMailValidator.class)
 @Target({ElementType.METHOD, ElementType.FIELD})
 @Retention(RetentionPolicy.RUNTIME)
-public @interface UniqueLogin {
-    String message() default "This login is already taken";
+public @interface UniqueMail {
+    String message() default "This mail is already registered";
     Class<?>[] groups() default {};
     Class<? extends Payload>[] payload() default {};
 }
